Module Ppx_decimal

A convenience syntax for writing decimal literals in source code, e.g. 1.1m. See the project readme for details on how to enable it.

val mk_of_string : loc:Ppxlib__.Import.location -> Ppxlib__.Import.expression

val expand : Ppxlib__.Import.location -> string -> Ppxlib__.Import.expression

val int_rule : Ppxlib.Context_free.Rule.t

val float_rule : Ppxlib.Context_free.Rule.t